chore(UI): move timeline to components [YTFRONT-4695]
