+++ packages/ui/src/ui/components/Timeline/index.ts
@@ -1,0 +1,2 @@
+export {Timeline} from './Timeline';
+export {YTTimeline, calcFromTo} from './YTTimeline';
